requirements.txt
attrs
schema
paho-mqtt